requirements-dev.txt
-r requirements.txt
pre-commit==1.12.0
PyHamcrest==1.9.0
pytest==3.9.1
